Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.1
pyyaml==6.0.1
ultralytics==8.0.227
aiofiles==23.2.1
scikit-learn==1.4.0
easyocr==1.7.1 